Change build scripts to work on Android.
Autoconf detects the presence of lutimes on Android, but older versions
of the API do not expose it publicly. This creates an issue for
compiling with -Werror, since C99 does not allow using a function
without declaring it.

Disabling -Werror does seem to allow it to build against lutimes, so it
would be possible to selectively ignore this warning, or to work around
it by providing our own prototype, but in the absence of the ability to
easily test multiple Android versions, this seems more robust.

Android API >= 26 provides lutimes in <sys/time.h>.

Termux on Android does not provide a user-writeable /tmp, but does
specify a temporary directory in TMPDIR. This will change the configure
script to use the TMPDIR environment variable if it is defined.

If TMPDIR is not defined, we default to /tmp.

bup can already be installed to a different prefix than the default of
/usr/local (also not writable in Termux) by specifying a different value
of PREFIX to make like so:

make PREFIX=/data/data/com.termux/files/usr install

Signed-off-by: Jamie Wyrick <terrifiedquack80@gmail.com>